feat: make Cursor.stream() work in pipeline mode
When sending the query from stream() in pipeline mode, we now use the
new pipeline_send() generator to also account for any command previously
in the pipeline. Results fetch is done through _stream_fetchone_gen()
method of Pipeline, which also possibly accounts for results of prior
commands. The single-row query is identified in the results queue by an
extra bool item in the tuple of PendingResult.

As can be seen in test_cursor_stream_query_fetch_bug(), there is a
problem when the pipeline contains a normal query after a streamed one,
apparently due to a bug in the libpq.

Cursor.stream() somehow interrupts the pipeline as it forces results
fetch for the single-row query, and thus for any query in the pipeline
queue (emitted before .stream()).