test: skip some tests that raise an error on pypy 3.9
The error is:

    anyio/_backends/_asyncio.py:2039: in _cancel_all_tasks
        raise cast(BaseException, task.exception())
    ...
    >               result = coro.throw(exc)
    E               RuntimeError: can't do async_generator.athrow().throw()

The error only happens on Pypy 3.9, and doesn't happen on master, only
in the maint 3.1 branch. 3.1 uses anyio < 4, master uses anyio >= 4, so
I would say that the error is limited to only this combination and that
it is an anyio bug.

The problem affects just two tests, so let's just skip them. Whatever.