fix(tstrings): disallow mixing of formats and sql objects
Force the t-string placeholders to have a format that specifies whether
the value must or must not be merged on the client side. So, if `:i` is
specified, the value must be either an Identifier or something that can
be transformed as such (a string).

Drop the idea of templates to be part of sql.SQL: all that we need is
that they can be an argument for SQL.join().