docs: this branch is now unsupported
psycopg-pool 3.2.x was maintained in this branch, but, because pool bug
fixes and releases are less frequent than changes to the rest of the
code, this branch has diverged hopelessly from maint-3.2, which is much
more modern.

It was way easier to rebase the pool changes onto the maint-3.2 branch
than to import the changes happened to the rest of the codebase in that
branch. Therefore the next psycopg-pool 3.2.x release will be tagged on
the maint-3.2 branch.